kun kall deleteAll mellomlagrede vedlegg hvis det finnes mellomlagrede vedlegg å slette

+++ src/main/kotlin/no/nav/sosialhjelp/soknad/vedlegg/fiks/MellomlagringService.kt
@@ -151,7 +151,10 @@ class MellomlagringService(
 
     fun deleteAllVedlegg(behandlingsId: String) {
         val navEksternId = if (MiljoUtils.isNonProduction()) createPrefixedBehandlingsId(behandlingsId) else behandlingsId
-        mellomlagringClient.deleteAllVedlegg(navEksternId = navEksternId)
+
+        if (getAllVedlegg(behandlingsId).isNotEmpty()) {
+            mellomlagringClient.deleteAllVedlegg(navEksternId = navEksternId)
+        }
     }
 
     companion object {
